Generalize compat check WIP
