+++ operators/pkg/controller/elasticsearch/elasticsearch_controller.go
@@ -25,15 +25,10 @@ import (
 	esreconcile "github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/reconcile"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/validation"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/utils/k8s"
-	semver "github.com/hashicorp/go-version"
-	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
-	v1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/tools/record"
-	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
@@ -195,7 +190,8 @@ func (r *ReconcileElasticsearch) Reconcile(request reconcile.Request) (reconcile
 		log.Info("Object is paused. Skipping reconciliation", "namespace", es.Namespace, "es_name", es.Name, "iteration", currentIteration)
 		return common.PauseRequeue, nil
 	}
-	compat, err := r.reconcileCompatibility(&es)
+
+	compat, err := annotation.ReconcileCompatibility(r.Client, &es, r.OperatorInfo.BuildInfo.Version)
 	if err != nil {
 		return reconcile.Result{}, err
 	}
@@ -294,76 +290,4 @@ func (r *ReconcileElasticsearch) finalizersFor(
 		keystore.Finalizer(k8s.ExtractNamespacedName(&es), r.dynamicWatches, "elasticsearch"),
 		http.DynamicWatchesFinalizer(r.dynamicWatches, es.Name, esname.ESNamer),
 	}
-}
-
-// reconcileCompatibility determines if this controller is compatible with a given resource by examining the controller version annotation
-// controller versions 0.9.0+ cannot reconcile resources created with earlier controllers, so this lets our controller skip those resources until they can be manually recreated
-// if an object does not have an annotation, it will determine if it is a new object or if it has been previously reconciled by an older controller version, as this annotation
-// was not applied by earlier controller versions. it will update the object's annotations indicating it is incompatible if so
-func (r *ReconcileElasticsearch) reconcileCompatibility(es *elasticsearchv1alpha1.Elasticsearch) (bool, error) {
-	annExists := es.Annotations != nil && es.Annotations[annotation.ControllerVersionAnnotation] != ""
-
-	// if the annotation does not exist, it might indicate it was reconciled by an older controller version that did not add the version annotation,
-	// in which case it is incompatible with the current controller, or it is a brand new resource that has not been reconciled by any controller yet
-	if !annExists {
-		exist, err := r.checkExistingResources(es)
-		if err != nil {
-			return false, err
-		}
-		if exist {
-			log.Info("Resource was previously reconciled by incompatible controller version and missing annotation, adding annotation", "controller_version", r.OperatorInfo.BuildInfo.Version, "namespace", es.Namespace, "es_name", es.Name)
-			err = annotation.UpdateControllerVersion(r.Client, es, "0.8.0-UNKNOWN")
-			return false, err
-		}
-		// no annotation exists and there are no existing resources, so this has not previously been reconciled
-		err = annotation.UpdateControllerVersion(r.Client, es, r.OperatorInfo.BuildInfo.Version)
-		return true, err
-	}
-
-	// if we have an annotation we need to check if it is compatible
-	currentVersion, err := semver.NewVersion(es.Annotations[annotation.ControllerVersionAnnotation])
-	if err != nil {
-		return false, errors.Wrap(err, "Error parsing current version on resource")
-	}
-	minVersion, err := semver.NewVersion("0.9.0-ALPHA")
-	if err != nil {
-		return false, errors.Wrap(err, "Error parsing minimum compatible version")
-	}
-	ctrlVersion, err := semver.NewVersion(r.OperatorInfo.BuildInfo.Version)
-	if err != nil {
-		return false, errors.Wrap(err, "Error parsing controller version")
-	}
-
-	// if the current version is gte the minimum version then they are compatible
-	if currentVersion.GreaterThanOrEqual(minVersion) {
-		log.V(1).Info("Current controller version on resource is compatible with running controller version", "controller_version", ctrlVersion.String(),
-			"resource_controller_version", currentVersion.String(), "namespace", es.Namespace, "es_name", es.Name)
-		return true, nil
-	}
-
-	log.Info("Resource was created with older version of operator, will not take action", "controller_version", ctrlVersion.String(),
-		"resource_controller_version", currentVersion.String(), "namespace", es.Namespace, "es_name", es.Name)
-	return false, nil
-}
-
-// checkExistingResources returns a bool indicating if there are existing resources created for a given resource
-func (r *ReconcileElasticsearch) checkExistingResources(es *elasticsearchv1alpha1.Elasticsearch) (bool, error) {
-	// if there's no controller version annotation on the ES instance, then we need to see maybe the CR has been reconciled by an older, incompatible controller version
-	selector := labels.Set(map[string]string{
-		label.ClusterNameLabelName: es.Name,
-	}).AsSelector()
-	opts := ctrlclient.ListOptions{
-		LabelSelector: selector,
-		Namespace:     es.Namespace,
-	}
-	var svcs v1.ServiceList
-	err := r.Client.List(&opts, &svcs)
-	if err != nil {
-		return false, err
-	}
-	// if we listed any services successfully, then we know this cluster was reconciled by an old version since any CRs reconciled by a 0.9.0+ operator would have a label
-	if len(svcs.Items) != 0 {
-		return true, err
-	}
-	return false, nil
 }
